fix(perf): spool-only hot path so CLI commands never load node:sqlite
recordSample appends NDJSON without importing sqlite.js. That keeps
ExperimentalWarning off stderr for ordinary commands (monitors --json tests
require empty stderr). SQLite opens only when aggregating (agents perf /
hooks profile).
