+++ apps/cli/src/index.ts
@@ -207,7 +207,6 @@ import type { AgentId } from './lib/types.js';
 import { IS_WINDOWS } from './lib/platform/index.js';
 import { getCliLaunch } from './lib/cli-entry.js';
 import { emit, emitFriction, redactArgs } from './lib/events.js';
-import { recordSample } from './lib/perf/db.js';
 
 // Transparent shim delegate: the generated Windows `.cmd` shims invoke
 // `agents __shim <agent>[@version] <raw args>`. Intercept here, before commander
@@ -296,14 +295,16 @@ program.hook('postAction', (_thisCommand, actionCommand) => {
       command,
       ...(durationMs !== undefined ? { durationMs } : {}),
     });
-    // Disposable perf warehouse — fail-soft, no FK to sessions.
+    // Disposable perf warehouse — fail-soft spool append (no SQLite on this path).
     if (durationMs !== undefined && parts[0] !== 'perf') {
-      recordSample({
-        kind: 'command.end',
-        label: command,
-        durationMs,
-        cwd: process.cwd(),
-      });
+      void import('./lib/perf/spool.js').then(({ recordSample }) => {
+        recordSample({
+          kind: 'command.end',
+          label: command,
+          durationMs,
+          cwd: process.cwd(),
+        });
+      }).catch(() => { /* fail soft */ });
     }
   } catch {
     // Best-effort completion record; the start line is the durable audit fact.
